fix: possible lossy conversion from double to long

+++ common/src/main/java/model/item/ServicePackage.java
@@ -6,8 +6,7 @@ package model.item;
  * Future extensions could include validity periods or service locations.
  */
 public class ServicePackage extends Item {
-
-    public ServicePackage(String id, String itemName, String description, double startingPrice) {
+    public ServicePackage(String id, String itemName, String description, long startingPrice) {
         super(id, itemName, description, startingPrice);
     }
